Fix nano size, change datatype to float

+++ code/virtualization/paging/measure-time.h
@@ -2,7 +2,7 @@
 
 struct  timespec    start, end;
 
-void            start_timer();
-void            end_timer();
-double          get_elapsed_seconds();
-unsigned long   get_elapsed_nano_seconds();
+void    start_timer();
+void    end_timer();
+float   get_elapsed_seconds();
+float   get_elapsed_nano_seconds();
